requirements.txt
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.103.2
h11==0.14.0
idna==3.10
jinja2==3.1.6
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.0.3
PyJWT==2.8.0
python-dotenv==0.21.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0
